+++ StateOfNeo/StateOfNeo.ViewModels/Address/AddressAssetViewModel.cs
@@ -8,6 +8,6 @@ namespace StateOfNeo.ViewModels.Address
     {
         public string Name { get; set; }
 
-        public decimal Balance { get; set; }
+        public float Balance { get; set; }
     }
 }
